+++ demos/session/src/main/java/androidx/media3/demo/session/MainActivity.kt
@@ -33,11 +33,12 @@ import android.widget.Toast
 import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
+import androidx.media3.cast.MediaRouteButtonFactory
 import androidx.media3.common.MediaItem
+import androidx.media3.common.util.UnstableApi
 import androidx.media3.session.LibraryResult
 import androidx.media3.session.MediaBrowser
 import androidx.media3.session.SessionToken
-import com.google.android.gms.cast.framework.CastButtonFactory
 import com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
 import com.google.common.util.concurrent.ListenableFuture
 
@@ -97,10 +98,11 @@ class MainActivity : AppCompatActivity() {
     }
   }
 
+  @OptIn(UnstableApi::class) // MediaRouteButtonFactory is unstable API.
   override fun onCreateOptionsMenu(menu: Menu): Boolean {
     super.onCreateOptionsMenu(menu)
     getMenuInflater().inflate(R.menu.menu, menu)
-    CastButtonFactory.setUpMediaRouteButton(this, menu, R.id.cast_menu_item)
+    val unused = MediaRouteButtonFactory.setUpMediaRouteButton(this, menu, R.id.cast_menu_item)
     return true
   }